From Functional Programming
On the first line you are given a list of numbers. On the next lines you are passed different commands that you need to apply to all numbers in the list: "add" -> adds 1; "multiply" -> multiplies by 2; "subtract" -> subtracts 1; "print" -> prints all numbers on a new line. The input will end with an "end" command, after which you need to print the result.