From Functional Programming
Write a predicate. Its goal is to check a name for its length and to return true if the names length is less or equal the passed integer. You will be given an integer that represents the length you have to use. On the second line you will be given a string array with some names. Print the names, passing the condition in the predicate.