From Functional Programming
Write a program which is using  a custom function (written by you) to find the smallest integer in a sequence of integers. The input could have more than one space. Your task is to collect the integers from the console, find the smallest one and print its index (if more than one such elements exist, print the index of the rightmost one).