From Defining Classes
You are the owner of a courier company and you want to make a system for tracking your cars and their cargo. Define a class Car that holds information about model, engine, cargo and a collection of exactly 4 tires. The engine, cargo and tire should be separate classes, create a constructor that receives all information about the Car and creates and initializes its inner components (engine, cargo and tires).
On the first line of the input you will receive a number N - the number of cars you have, on each of the next N lines you will receive information about a car in the format "<Model> <EngineSpeed> <EnginePower> <CargoWeight> <CargoType> <Tire1Pressure> <Tire1Age> <Tire2Pressure> <Tire2Age> <Tire3Pressure> <Tire3Age> <Tire4Pressure> <Tire4Age>" where the speed, power, weight and tire age are integers, tire pressure is a double. 
After the N lines you will receive a single line with one of 2 commands "fragile" or "flamable" , if the command is "fragile" print all cars whose Cargo Type is "fragile" with a tire whose pressure is  < 1, if the command is "flamable" print all cars whose Cargo Type is "flamable" and have Engine Power > 250. The cars should be printed in order of appearing in the input.