Generic Swap Method Strings
Create a generic method that receives a list containing any type of data and swaps the elements at two given indexes.
As in the previous problems, read n number of boxes of type String and add them to the list. On the next line, however, you will receive a swap command consisting of two indexes. Use the method you've created to swap the elements that correspond to the given indexes and print each element in the list.