Generic Count Method Strings
Create a method that receives as argument a list of any type that can be compared and an element of the given type. The method should return the count of elements that are greater than the value of the given element. Modify your Box class to support comparing by value of the data stored. 
On the first line you will receive n - the number of elements to add to the list. On the next n lines, you will receive the actual elements. On the last line you will get the value of the element to which you need to compare every element in the list.